preview cleanup
